Changed data type in matrix from int to double

+++ Matrix.h
@@ -7,7 +7,7 @@
 class Matrix {
 private:
 	size_t H, W, N; // H - height (rows), W - width (columns), N - total number of elements (H * W)
-	int** data;
+	double** data;
 
 public:
 
@@ -24,12 +24,12 @@ public:
 	Matrix operator - (Matrix& A);    // Matrix substraction
 	Matrix operator * (Matrix& A);    // Matrix multiplication
 	
-	int operator ()(size_t h, size_t w);  // Returns a single element
-	void setElement(size_t h, size_t w, int value);  // Assigns the value to the data[h][w]
+	double operator ()(size_t h, size_t w);  // Returns a single element
+	void setElement(size_t h, size_t w, double value);  // Assigns the value to the data[h][w]
 
-	Matrix getMultipliedBy(int scalar);   // Returns matrix multiplied by a scalar
+	Matrix getMultipliedBy(double scalar);   // Returns matrix multiplied by a scalar
 	Matrix getRaisedTo(int power);   // Returns matrix raised into power
-	long long calculateTrace();    // Calculates and returns a trace of the matrix
+	double calculateTrace();    // Calculates and returns a trace of the matrix
 	double calculateDeterminant(); // Calculates and returns a determinant of the matrix
 	Matrix getMinor(size_t row, size_t col); // Returns a minor by excluding one row and one column from the original matrix
 	Matrix getTransposed();  // Returns a transposed matrix
